2004-01-21  Elliott Hughes  <enh@acm.org>
	* src/body.cpp (ProcessTryStatement): Better error report on
	duplicate declaration.
	* src/error.cpp (INVALID_CHARACTER_VALUE),
	(INVALID_CAST_CONVERSION): Fix typos.
	* src/error.h: Ditto.
	* src/lookup.h: Ditto.
	* src/semantic.h (ProcessClassFile): Fix prototype.

+++ jikes/src/semantic.h
@@ -1308,7 +1308,7 @@ private:
                                     LexStream::TokenIndex);
     TypeSymbol* GetType(TypeSymbol*, CPClassInfo*, const ConstantPool&,
                         LexStream::TokenIndex);
-    void ProcessClassFile(TypeSymbol*, const char*, size_t,
+    void ProcessClassFile(TypeSymbol*, const char*, unsigned,
                           LexStream::TokenIndex);
     void ReadClassFile(TypeSymbol*, LexStream::TokenIndex);
